booking/requirements.txt
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
defusedxml==0.7.1
Django==6.0.2
django-allauth==0.50.0
djangorestframework==3.16.1
idna==3.11
oauthlib==3.3.1
pillow==12.1.0
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.5
urllib3==2.6.3
